+++ src/main/java/com/backend/tienda/entidades/Clientes.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 
 public class Clientes {
     @Id
-    private Long cedula_cliente;
+    private Double cedula_cliente;
 
     private String direccion_cliente;
     private String email_cliente;
