+++ src/main/java/com/backend/tienda/repository/DetalleVentasRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.DetalleVentas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleVentasRepository extends JpaRepository<DetalleVentas, Long> {
+public interface DetalleVentasRepository extends JpaRepository<DetalleVentas, Double> {
 }
