+++ src/main/java/com/backend/tienda/repository/ProductosRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.Productos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductosRepository extends JpaRepository<Productos, Long> {
+public interface ProductosRepository extends JpaRepository<Productos, Double> {
 }
