+++ src/main/java/com/backend/tienda/repository/VentasRepository.java
@@ -3,5 +3,5 @@ package com.backend.tienda.repository;
 import com.backend.tienda.entidades.Ventas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentasRepository extends JpaRepository<Ventas, Long> {
+public interface VentasRepository extends JpaRepository<Ventas, Double> {
 }
