+++ src/main/java/com/backend/tienda/servicios/ProductosServicios.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ProductosServicios {
     public Productos actualizarProductos(Productos productos);
-    public void eliminarProductos(Long codigo_producto);
+    public void eliminarProductos(Double codigo_producto);
     public Productos guardarProductos(Productos productos);
     public List<Productos> listarProductos();
-    public Productos buscarProducto(Long codigo_producto);
+    public Productos buscarProducto(Double codigo_producto);
 }
